change js structure
