+++ form.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quiz App</title>
     <link rel="stylesheet" href="css/global.css" />
-    <script type="text/javascript" src="js/form.js" defer></script>
+    <script type="module" src="./js/components/CreateNewCard.js" defer></script>
   </head>
   <body data-js="body">
     <header class="header" data-js="header">
